Full tests refactor
